Change decimal to integer values in the migration file

+++ db/migrate/20231024182658_create_expenses.rb
@@ -2,7 +2,7 @@ class CreateExpenses < ActiveRecord::Migration[7.1]
   def change
     create_table :expenses do |t|
       t.string :name
-      t.decimal :amount
+      t.integer :amount
       t.references :author, null: false, foreign_key: { to_table: :users }
       t.references :category, null: false, foreign_key: true
